feat(addressindex): proper P2TR support via uint256 key schema
  The address index inherited from a pre-taproot codebase used uint160
  throughout, silently dropping P2TR outputs since a 32-byte x-only
  pubkey can't fit in 20 bytes. With taproot always-active from genesis
  and no mainnet data yet, this is the right time to fix the schema
  correctly rather than retrofit around it.

  All key structures (CAddressIndexKey, CAddressUnspentKey, iterator
  keys, CMempoolAddressDeltaKey) now use uint256. Legacy address types
  (P2PKH, P2SH, P2WPKH) zero-pad their 20-byte hash into the low bytes
  of the 32-byte field — LevelDB range scan ordering is preserved. P2TR
  (address type 4) stores the full x-only witness program directly,
  allowing getAddressFromIndex to reconstruct a real bech32m address via
  XOnlyPubKey → WitnessV1Taproot → EncodeDestination. All getaddress*
  RPCs now work correctly with taproot addresses end-to-end.

  This change only affects nodes running -addressindex (off by default).
  Any such node upgrading from a previous index build will need to
  delete indexes/addressindex/ and resync.

+++ src/index/addressindex.h
@@ -48,17 +48,17 @@ public:
     virtual ~AddressIndex() override;
 
     /// Read all deltas for an address (optionally filtered by asset and/or block range).
-    bool ReadAddressIndex(uint160 addressHash, int type, std::string assetName,
+    bool ReadAddressIndex(uint256 addressHash, int type, std::string assetName,
                           std::vector<std::pair<CAddressIndexKey, CAmount>>& addressIndex,
                           int start = 0, int end = 0);
-    bool ReadAddressIndex(uint160 addressHash, int type,
+    bool ReadAddressIndex(uint256 addressHash, int type,
                           std::vector<std::pair<CAddressIndexKey, CAmount>>& addressIndex,
                           int start = 0, int end = 0);
 
     /// Read unspent outputs for an address (optionally filtered by asset).
-    bool ReadAddressUnspentIndex(uint160 addressHash, int type, std::string assetName,
+    bool ReadAddressUnspentIndex(uint256 addressHash, int type, std::string assetName,
                                  std::vector<std::pair<CAddressUnspentKey, CAddressUnspentValue>>& unspentOutputs);
-    bool ReadAddressUnspentIndex(uint160 addressHash, int type,
+    bool ReadAddressUnspentIndex(uint256 addressHash, int type,
                                  std::vector<std::pair<CAddressUnspentKey, CAddressUnspentValue>>& unspentOutputs);
 };
